+++ dhis-2/dhis-web/dhis-web-maintenance/dhis-web-maintenance-mobile/src/main/java/org/hisp/dhis/mobile/action/DeleteSentSMSAction.java
@@ -74,14 +74,14 @@ public class DeleteSentSMSAction implements Action
     {
         if ( ids != null && ids.length > 0 )
         {
-            for ( Integer each : ids )
+            for ( long each : ids )
             {
                 outboundSmsService.deleteById( each );
             }
         }
         if ( id != null )
         {
-            outboundSmsService.deleteById( id );
+            outboundSmsService.deleteById( (long)id );
         }
         return SUCCESS;
